tweak schema build so its more modular and easier to port.
generally make work with oracle

add READMEs that explain files and how to build things

packages don't all build yet; not quite sure wtf is going on in some cases,
but they're less important, especially the VOE one that fails because it
needs to be rearchitected.

some triggers need to be revisited.