Added 'order' parameter to DBFetch to specify a key or keys to order by, which is passed directoy to the query as "ORDER BY ...".  the key is free-form, so to specify a descendant ordering, you have to say "key desc".  This may or may not change.  The value of the order parameter must be either a scalar or an array reference.
Fixed GetNetblock to treat an undefined netmask_bits parameter to return
all of the netblocks with a matching ip_address, ordered by longest
subnet mask.

Allowed 'single' parameter to GetNetblock to support a 'first' value which
does not error if multiple rows are returned to the query, but will
return the first value without erroring.  There is no way to determine
whether multiple values were returned; if you want that, omit 'single' and
look at the array reference returned yourself.