Added additional capabilities to netblock_utils.find_free_netblock(s) and with some parameter extensions to netblock_manip.allocate_netblock.
The allocate_from_bottom parameter was removed and replaced with a new
parameter 'allocate_method' which takes one of four values: 'bottom', 'top',
'random', and 'default'

'bottom' and 'top' represent the previous values for 'allocate_from_bottom'
being 'true' or 'false' respectively.

'random' is only valid for single addresses, where a random free netblock
from the list of parent netblocks is returned.

'bottom' is the default for all cases except allocating single IPv6 addresses
from a netblock at least as large as a /110, in which case 'random' is
the default.

Due to the way 'random' is implemented, it is possible that in a sufficiently
congested netblock of appropriate size (on the order of an IPv4 /22
or IPv6 /118), using the default rnd_max_count may result in an address that
is actually free in the netblock not being found to be assigned.  In
practice, the default values will never have this issue before the
heat death of the universe.

Additionally, an 'ip_address' parameter may be passed to allocate_netblock
or 'desired_ip_address' to fine_free_netblock(s) to attempt to locate/allocate
a netblock with a specific ip_address located within a parent or set of
parents.