Make layerx_network_manip.delete_layer2_networks call delete_layer3_networks for handling the layer3 things.
Fix issues with deleting layer3_network components that are parts of shared_netblocks
or are assigned to network_interfaces with multiple addresses.  Also handle purging
netblocks that are part of netblock_collections correctly.

Add better errors to netblock triggers.  This still need more overhaul in someone's
copious spare time, but then, so do lots of errors.